test eth_to_gear relayer
