+++ tests/Cargo.toml
@@ -4,6 +4,7 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+async-trait.workspace = true
 futures.workspace = true
 gclient.workspace = true
 gear-core.workspace = true
